commonizing cli logic
